app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
openai
dateparser
passlib[bcrypt]
python-jose[cryptography]
python-dateutil
psycopg2-binary
